[docs] Components concepts page (#29017)
## Summary & Motivation

Add components to the concepts page. Trying not to overcomplicate things
too much and just show that a `definitions` is the output of a
`component`. Updating the main diagram, definitions diagram and new
components diagram.

**Main Diagram**
![Screenshot 2025-04-04 at 12 22
40 PM](https://github.com/user-attachments/assets/fd742772-0686-4e7a-b991-dd5fdd3ae2a3)

Don't love that the arrows go to components box rather than the
definition but the mermaid diagram looks really weird when pointing to
the definitions.

**Component Diagram**
![Screenshot 2025-04-04 at 12 22
27 PM](https://github.com/user-attachments/assets/f8975ae9-6be8-4c21-ba46-555bd7be88d8)

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.