[asset health] derive health from latest asset observation for observable assets (#29322)
## Summary & Motivation
Some assets are only ever observed, which means the current health
reporting component for materialization status always shows them as
UNKNOWN. This adds a case to handle observable assets. If they have an
observation they are healthy.

We don't have a mechanism for keeping track of observation failure (like
if the compute function fails) so if an asset successfully observes,
then fails, it still shows as healthy. tbd if it's worth investing the
effort to make failure events for this...

Assets that are materializable and observable always use latest
materialization status when computing health

## How I Tested These Changes

tests in https://github.com/dagster-io/internal/pull/15127