[resolved] Propagate default value to inner ResolvableModel schema when possible (#29315)
## Summary & Motivation

This prevents the `__DAGSTER_UNSET_DEFAULT__` thing from showing up all over the docs, I believe. Also good for tooling in general, as it makes the default values get represented in the json schema where possible.

Sometimes, the default value will not be serializble (e.g. it's a default factory or a complex object type), and then we fall back to the unsightly string

## How I Tested These Changes

## Changelog

NOCHANGELOG