[ui] Configurable "faceted" asset graph node designs (#29307)
## Summary & Motivation

Linear:
https://linear.app/dagster-labs/issue/OPER-1793/configurable-asset-graph-node-designs-prototype

Figma:
https://www.figma.com/design/wXU4we6PavPVZt5dHZNQkY/Dagster-Observe---2025?node-id=1014-814&p=f&t=BY6WnhP5GaXjOz7F-0

This PR adds a new feature flag which enables faceted asset node
rendering. In this new rendering style, you can choose exactly what bits
of information you'd like to see on the asset graph, and data is
presented in a more uniform way. All the nodes are the same size, and
most facets are rendered in a `[key] [value]` style within the node box.

After this merges, I plan to follow-up with:

- Updated design for the bottom `status` bar which will use the new
asset health data (subject to a separate feature flag)
- An additional "Automation condition" facet which will use the new
automation condition data

One noteworthy / non-flagged change in this PR is that layout updates
that do not result in a change in the number of edges, groups or nodes
do NOT re-center the graph. This allows the user to change their
displayed facets without losing their place in the graph.

Sidenote: Yesterday I mentioned I was trying to dynamically "tweak" the
graph layouts to change node heights without re-running the layout
algorithm. I was 1) unable to do this because external asset nodes are
still tiny and do not scale, and scaling was creating odd spaces around
them and 2) not seeing changes to asset height impact the placement of
nodes in most cases anyway.


![image](https://github.com/user-attachments/assets/88793afa-88d2-4c37-84e8-d1135cc8eede)

![image](https://github.com/user-attachments/assets/e52288d1-406f-4c96-beee-c4d0b8a4b2ed)


![image](https://github.com/user-attachments/assets/a8554157-bc1a-4228-864b-32c355028f98)

![image](https://github.com/user-attachments/assets/898bbc84-b800-45bc-8f07-e2b513dfd040)

## How I Tested These Changes

Tested feature flag on and off on both the asset graph and the asset
lineage view, and also zoomed out to test minimal rendering.

## Changelog

[ui] A new feature flag allows you to customize the appearance of assets
on the asset graph by enabling and disabling individual facets.

---------

Co-authored-by: bengotow <bgotow@elementl.com>