[dg] Add `dg scaffold github-action` command support for hybrid orgs (#29065)
## Summary

Stacks on #28962. Implements `dg scaffold github-action` for hybrid
orgs.

`dg scaffold github-action` queries Plus to figure out which agent type
is in use, and scaffolds the corresponding GitHub actions file. It looks
for a `regsitry.yaml` file to determine which registry type to log in
to, and scaffolds a build step and dockerfile (if one doesn't yet exist)
for each location.


## Test Plan

Mocked unit tests.