change UI to narrow asset events query (#28903)
## Summary & Motivation
This PR simplifies the data fetching for metadata plots in a couple
ways.

Before:
For partitioned assets, we were using the `partitionInLast` parameter to
filter materializations. This fetches all partition keys (expensive),
and then grabs the last N materializations that show up in the last N
partitions* (incorrect). We were also fetching the materialization in
the most recent partition to represent the most recent materialization
overall.

* This incorrectness is better documented here:
https://github.com/dagster-io/dagster/pull/25696

After:
We separate out the most recent materialization query to be just that,
fetch the most recent materialization regardless of partition (this may
or may not be the right decision, based on product, but it's easy to
change). For partitioned assets, we now do a two-phased fetch. We first
fetch the last N partition keys, using the new paginated partition key
graphql fields. We then use those partition keys to pass into the
`latestMaterializationByPartition` field, which gets the most recent
materialization for the provided set of partition keys.

Breaking apart the partition history data fetching to be closer to the
graphs component also means that the main components become interactive
much earlier.

This should result in better correctness and massively reduce
overfetching. This should also result in much better performance in
Cloud.

## How I Tested These Changes


https://github.com/user-attachments/assets/04a37bc2-a4c5-4d06-80b9-549510767e9d



## Changelog