[gql] latestEventSortKey resolver for GrapheneAsset so that we can order by id (#29264)
## Summary & Motivation
resolver on `GrapheneAsset` that returns the latest storage id of any
event associated with that asset. This lets us sort assets by the time
they were last "modified" (planned, materialized, failed, observed).

Figured i'd do storage id since its 1) monotonically increasing and 2)
already available for planned events on AssetEntry (we don't store the
full event or the event timestamp for planned events)


Some perf stats pulled from shadow-gql against `dogfood-test-1` as a
place to start:
`AssetCatalogTableQuery` with limit of 10000 as used in the UI (which is
effectively fetching all assets in dogfood-test-1)
`master`: 43s, 37s, 37s
this branch: 39s, 42s, 38s

`AssetCatalogTableQuery` with limit of 10000 times as reported in the
network tab:
`dogfood-test-1`: 9.6s, 8.2s, 9.3s
`elementl` prod: 6.3s, 4.9s, 8.8s
documenting these if we want to compare after this branch merges and the
`AssetCatalogTableQuery` is updated, but it'll be easier to do more
direct comparison of the `AssetCatalogTableQuery` with and without this
new resolver once it lands


## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.