Remove asset check run observations (#29453)
## Summary & Motivation

https://github.com/dagster-io/dagster/pull/16290 added logic to observe
the latest run for asset checks but causes an issue where we observe
runs regardless of whether they're inprogress. To do that we open a
websocket connection for every unique run id of every asset check of
every visible asset. I believe the intention was to only observe runs
that are in progress but currently we have no way of knowing that for
asset checks and end up observing the latest run regardless of its
state.

## Test plan

There are no existing jest tests for this... I can look into that if we
want to re-enable this behavior.