[dg] Rebuild cache on deserialization error (BUILD-997) (#29434)
## Summary & Motivation

Currently `dg` chokes with an inscrutable error when it reads from a
cache in an outdated data format (causing deserialization errors). This
PR changes the behavior so that deserialization errors are caught and
treated the same as a missing cache entry (a message is printed saying
the cache is outdated or invalidated).

## How I Tested These Changes

New unit tests.

## Changelog

Cache deserialization errors are now ignored. Previously, when the `dg`
contents of the cache were in an outdated format, `dg` could crash. Now
it will just rebuild the cache.