Wrap AssetView with key (#29470)
## Summary & Motivation

We got a report that the materialize button on the asset page was
materializing a previously visited page. This suggests that we're
reusing the <LaunchAssetExecutionButton/> in the page header and likely
other elements on the page. To prevent this issue lets set a key so that
we don't re-use the component instance.