[dagster-dg] Add separate command to generate JSON schema for code location (#27896)
## Summary

Adds a command to generate a JSON schema for a given code location.
Drops the schema in the location referenced by the VS Code extension, so
this can be used to update the schema without the other rigamarole
involved with that command.

```sh
dg code-location generate-component-schema --output-file foo.json
```

When #27863 lands, I intend to add a `--watch` flag to this command as
well, so you can live-update the schema if the set of components
changes.

## Test Plan

Brief test, the underlying `dagster-components` CLI tests correctness.