[resolved] split asset defs update and asset spec update handling (#29490)
Currently we just have `AssetAttributes` which we use to power both `Definitions` post processing as well as pre-definitions `AssetSpec` customization. 

Unfortunately, these two context have different constraints. We do not support changing the key of an asset spec inside of an already formed `AssetsDefinition` as its difficult to ensure it could safely handle that transformation in all cases. 

This change separates the two cases, and adds `key_prefix` for the spec only update condition. 

## How I Tested These Changes

added tests