[ui] Support new freshness row on faceted asset nodes (#29520)
This is a small update that improves support for "Freshness Policy" on
the new faceted asset nodes.

This row supports the old freshness policy concept all the time, and the
new one if the `Observe UI` feature flag is enabled and a new freshness
policy is present. (It's valid to have Observe UI turned on and an old
style freshness policy)



Example of an asset with an old freshness policy:


![image](https://github.com/user-attachments/assets/8a8d58c9-1edc-42d9-8edc-c2a265c18aa1)

![image](https://github.com/user-attachments/assets/7e45930d-0ecc-48e3-b078-bc83c0088152)

Example of an asset with a new freshness policy and the Observe UI
feature enabled:


![image](https://github.com/user-attachments/assets/294ca3bf-019c-45b8-9a7e-1eb10e840bd1)

![image](https://github.com/user-attachments/assets/d2c3376f-267b-422d-9a32-32f5319bd107)

![image](https://github.com/user-attachments/assets/828bdb44-c353-4447-9809-ce34bda92f0b)

## Testing:

I verified these on live assets and also in the storybook 


![image](https://github.com/user-attachments/assets/2060819e-c9fd-4d17-86fd-fae6a719aa77)

Co-authored-by: bengotow <bgotow@elementl.com>