[dg] Update dg list plugin IPC (#29513)
## Summary & Motivation

- Change `dagster-components list library` -> `list plugins`. This
should have been done when the `dg` command was converted.
- The output format has been changed. `list libraries` outputs a
`list[PluginObjectSnap]`. `list plugins` outputs a `PluginManifest`. The
`PluginManifest` contains a `list[PluginObjectSnap]` but also a list of
all modules scanned. This means it contains information about which
plugins were scanned but contained no objects. This is useful to detect
the situation where a project declares an entry point but it isn't
registered (see
[upstack](https://github.com/dagster-io/dagster/pull/29514)).
- `dg` will detect when it's dealing with a version with `list plugins`
or `list libraries` and call the appropriate command. It also normalizes
the differing output formats to the same `PluginManifest` format on the
`dg` end.

## How I Tested These Changes

New unit tests, including backcompat tests.