Add `operator_type` to AutomationConditionNodeSnapshot (#29502)
## Summary & Motivation

This will be used to provide the UI with more semantic information about how a given condition works.

To start, we'll use this to help determine the collapsed state of each node in the evaluation tree, but we may also end up using this for other parts of the rendering to make it more obvious how results are being combined together.

## How I Tested These Changes

## Changelog

NOCHANGELOG