[da] Allow AutomationResult to take a `metadata` parameter (#29509)
## Summary & Motivation

This is used upstack to add additional metadata to the evaluation of a SinceCondition.

This likely has other applications as well (e.g. the EntityMatches condition could store AssetKeyMetadataValues to be a bit more explicit about what asset is getting evaluated).

While it's a bit funky to use SubsetsWithMetadata for this sort of thing, I deemed this safer than adding a whole extra parallel storage field. It's sorta nice to just reuse what we have here.

## How I Tested These Changes

## Changelog

NOCHANGELOG