dagster-k8s: handle jobs with backoff retries (#29380)
## Summary & Motivation
We use karpenter as our k8s autoscaler. Karpenter's handling of
daemonsets sometimes leads to pod preemption in k8s (see their
[FAQ](https://karpenter.sh/docs/faq/#when-deploying-an-additional-daemonset-to-my-cluster-why-does-karpenter-not-scale-up-my-nodes-to-support-the-extra-daemonset)
for reference). As such, we increased the backoff_limit on our dagster
jobs using the K8sRunLauncher.

Unfortunately, this is leading to false failure detection in k8s run
monitoring. Here's an example of a dagster job that was declared failed.
```
spec:
  backoffLimit: 2
  completionMode: NonIndexed
  completions: 1
  manualSelector: false
  parallelism: 1
  podReplacementPolicy: TerminatingOrFailed
...
status:
  active: 1
  failed: 1
  ready: 1
  startTime: "2025-04-17T19:52:23Z"
  terminating: 0
  uncountedTerminatedPods: {}
```

k8s event logs show that the original pod was preempted, so it was
rescheduled.
```
4m10s       Normal    SuccessfulCreate       job/dagster-run-269c122e-9f18-42bc-b9de-7a8047578b91         Created pod: dagster-run-269c122e-9f18-42bc-b9de-7a8047578b91-d8qg8
4m20s       Normal    Scheduled              pod/dagster-run-269c122e-9f18-42bc-b9de-7a8047578b91-ttjmx   Successfully assigned dagster/dagster-run-269c122e-9f18-42bc-b9de-7a8047578b91-ttjmx to ip-172-24-8-32.ec2.internal
4m19s       Normal    Preempted              pod/dagster-run-269c122e-9f18-42bc-b9de-7a8047578b91-ttjmx   Preempted by pod 9cda2b57-b855-4e1d-a66a-d4e506b96416 on node ip-172-24-8-32.ec2.internal

4m10s       Normal    SuccessfulCreate       job/dagster-run-269c122e-9f18-42bc-b9de-7a8047578b91         Created pod: dagster-run-269c122e-9f18-42bc-b9de-7a8047578b91-d8qg8
```

In other words, the original pod was preempted before the job even
started, and then a new pod was launched. In dagster UI, we get the
following:

![image](https://github.com/user-attachments/assets/79c17d6e-970e-4f69-9eca-e323ed282ab7)
The job starts running at this point, but later a "failure" is recorded
since run monitoring doesn't handle the backoff.

![image](https://github.com/user-attachments/assets/d88d27f4-0814-4711-9838-cf712320dd99)

I'm very open to other solutions here, but thought it'd be more
effective to start a discussion with a proposed solution.

This builds on work from #16303 and may resolve (at least part of)
#6236.

## How I Tested These Changes
Added to unit tests

## Changelog

dagster-k8s: improved run monitoring when running with increased backoff
limits

---------

Co-authored-by: Deepyaman Datta <deepyaman.datta@utexas.edu>