[dg] Warn on outdated dagster-dg (BUILD-998) (#29493)
## Summary & Motivation

Warns when a newer version of `dagster-dg` is available.

- If the cache is enabled, then PyPI will only be hit once per hour.
- The check can be prevented from ever running by suppressing the
`dg_outdated` warning.
- A special private env var `DAGSTER_DG_UPDATE_CHECK_ENABLED` is also
checked. We set this to 0 in tests to prevent bombing the PyPI API in
CI.

The warning looks like this:

```
There is a new version of `dagster-dg` available:

    Latest version: 0.26.11
    Installed version: 0.26.11rc1

Update your dagster-dg installation to keep up to date:

    [uv tool]  $ uv tool upgrade dagster-dg
    [uv local] $ uv sync --upgrade dagster-dg
    [pip]      $ pip install --upgrade dagster-dg

To suppress this warning, add "dg_outdated" to the `cli.suppress_warnings` list in your configuration.
```

## How I Tested These Changes

New unit tests.

## Changelog

`dg` now checks for new versions and prompts the user to update. The
check runs automatically at `dg` startup once per day.