[dg] Fail on dagster-dg version less than dagster version (BUILD-1041) (#29505)
## Summary & Motivation

Exit with an error when the `dg` version is less than the dagster in the
in-scope environment. We have to transform the "core version" of
`dagster` to a "library version" in order to perform the comparison.

The check is run whenever we spawn a `dagster-components` subprocess.
## How I Tested These Changes

New unit tests.

## Changelog

`dg` will now fail with an error message if it's version is below the
minimum supported version for the version of `dagster` in your
environment.

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>