Refactor workspace context (#29592)
## Summary & Motivation

Refactor WorkspaceContext with a couple of goals in mind:

1. (Main goal) Extract the core location based fetching logic into its
own abstract class `BaseDataFetcher` (Probably needs a better name) so
that I can re-use it for fetching assets by location in a follow up PR.
2. Extract the majority of logic outside of React and instead keep a
thin layer in the provider. This removes the need to use effects /
manage dependency arrays which makes the logic easier to understand.
3. Removed the refetch function which was only called by
`MergedAutomationRoot` for an invalid reason (it was being called when
an automation starts/stops but that data is cached by Apollo a). We
should never need to manually refetch the workspace context since we're
polling code location status.

## How I Tested These Changes

1. All of the existing WorkspaceContext tests pass.
2. New jest tests for new classes
3. Used proxy to load a variety of orgs.
4. Local host cloud tested adding a code location