Replace calls to PersistedTimeWindow.start.timestamp() and PersistedTimeWindow.end.timestamp() with start_timestamp and end_timestamp properties (#29599)
## Summary & Motivation
PersistedTimeWindow starts with a UTC timestamp, so transforming this into a datetime just to transform it back into a UTC timestamp is inefficient. Furthermore we have some evidence of lock contention in cached_property calls in python versions < 3.12, so we can avoid using a cached_property in a hot path that we don't need.

The main unfortunate thing here is that there are callsites that mix the user-facing TimeWindow class (which does not have a start_timestamp or end_timestamp property) with this PersistedTimeWindow class that does. Typechecking should help us avoid errors here but we could consider adding the properties to TimeWindow as well as a backstop.

## How I Tested These Changes
BK