[1/2] Add a resolver to get all materialized asset keys (#29626)
## Summary & Motivation

Adds a resolver to fetch all materialized asset keys without also
fetching their definitions. This uses a new resolver instead of
`assetsOrError` in order to avoid fetching the entire asset graph. I
took a bit of a shortcut (sorry @alangenfeld) in that I'm just returning
the asset keys for now but in the future we might want to add more
fields to `GrapheneAssetState` that can be supported solely by the DB
row in which case we'd need to add new storage methods (using existing
logic) then.

In the next PR in this sequence we're updating our asset fetching to go
through our WorkspaceContext on the frontend meaning we end up going
through `workspaceLocationEntryOrError`. The only issue is that assets
without a definition aren't returned by that resolver since they're not
defined in the workspace so we still need to fetch all materialized
assets to know about external assets hence this PR.

## How I Tested These Changes
Snapshot test + Manual testing with frontend PR
https://github.com/dagster-io/dagster/pull/29627.