[2/2] Fetch assets as part of workspace query (#29627)
## Summary & Motivation
Motivation: Perf.

We want to move away from the `assetsOrError` resolver because it
fetches the entire asset graph and is fairly slow. Instead we're going
to use `assetsStateOrError` resolver which avoids fetching the entire
asset graph and only returns the asset keys (much much faster). To get
the definitions we rely on the WorkspaceContext which will now include
asset definitions for each location. Then we combine the two sources in
order to get the full list of SDAs + external assets. Note that we don't
paginate the fetching in the WorkspaceContext but this is fine since
everything is already loaded in memory (from the workspace snapshot) and
in testing with large orgs the impact on perf was negligible.

Perf. benefits:

- Less data being loaded from / saved to indexeddb on every load (and
every 60 seconds)
- SDA asset fetching is parallelized by code location and only
re-fetched if it changes.

## How I Tested These Changes

Updated existing jest tests.
Manual testing with local host cloud with special attention to
performance.

I still need to check the case of when an asset is defined in multiple
code locations and re-create the tie breaking strategy we use on the
backend.