[da] Fix initial_evaluation condition handling of missing automation conditions (BUILD-684) (#27630)
## Summary & Motivation

This fixes a bug as described in the changelog entry.

The existing logic for updating the AssetDaemonCursor ensured that we would never "forget" about the cursor of any asset that was ever evaluated by a given process.

This is a bit of a holdover from the world of AMP + global AMP evaluation in which:

1. it was possible for a single code-location to be unavailable while the rest of the evaluation was proceeding normally
2. it was very bad (perf-wise) if an AMP cursor was lost to the void

So the existing code was the safest possible measure -- we just literally always propagate old cursors forward. This means that over time, if you change asset names or change how things are automated, your cursor just hold on to a lot of baggage.

This PR makes a small step in the direction of clearing out this baggage, but deeming it safe to remove the cursor of an asset that is still in the asset graph (so we know about it), but did not have any cursoring information (meaning it was explicitly not being automated by DA)

So we still keep around some amount of baggage (to continue to support the scary situation from AMP-times), but get to remove some subset of it.

My belief is that once we remove the "global" mode of the daemon, we should be able to remove all of the propagation logic, but I haven't thought about that too hard.

## How I Tested These Changes

## Changelog

Fixed a bug with `AutomationCondition.initial_evaluation` which could cause it to return `False` for an asset that went from having a condition, to having no condition at all, back to having the original condition again.