[ui] Support all markdown features used in integration marketplace pages (#29622)
This PR addresses
https://linear.app/dagster-labs/issue/BUILD-980/dg-in-app-marketplace-render-all-integration-markdown-correctly

There’s a block comment in MarkdownSupport.tsx explaining the approach,
but this fixes rendering of:

- `<Beta/>` and `<Deprecated />`
- tables and lists (CSS tweaks)
- code blocks containing yaml
- tabs
- prerequisites sections
- directives like `:::note`
- `<CodeExample />` which was not always inlined due to strict regexp
matching of props
- `<PyObject />` references

It’s worth noting that we won’t get support for new inline React
elements for free - each custom JSX element used in the integration docs
needs to be implemented in the app as well - given the low number of
components and the drastically different environments, I think this is
an ok tradeoff!

To test this, I updated the Storybooks to include two real integration
markdown files, sling and fivetran, which together use all of the
features above.

Co-authored-by: bengotow <bgotow@elementl.com>