Expose asset_selection param to submit_job_execution function.
Modify dagster_graphql client function `submit_job_execution` to accept a list of
CoercibleToAssetKey in a parameter called `asset_selection`. This is then forwarded
to the GraphQL corresponding function to correctly submit jobs for a given subset
of assets when this parameter is passed.

A new test was added to ensure the parameter is being passed correctly.