[historical-runs 2/n] Add history import run method (#29381)
## Summary & Motivation
Adds a add_historical_run method to run storage which can be used to
alter the create_timestamp of the underlying run. This is necessary to
import historical runs and have them represent properly in any views
built/filtered on create_timestamp in some way.

## How I Tested These Changes
Added new storage tests for the method.