[historical-runs 1/n][dagster-airlift] make monitor job range configurable (#29339)
## Summary & Motivation
Switches the monitor job to operate over a range specified via config.
This opens the door to make it usable for history import.

The upshot is that we switch to a sensor instead of a schedule for
kicking off compute so that we can manage the range interval we're
operating over in a cursor. I think this is actually nicer / more
observable overall.

## How I Tested These Changes
New tests for sensor behavior, updated existing monitor job tests.