[dagster-airlift][print metadata 2/2] Ingest metadata from airflow logs (#29313)
## Summary & Motivation
Hooks the metadata APIs up from previous PRs into the monitoring job; so
that we actually log the external metadata on the object.

The task logs API is per-task, and so we parallelize using asyncio to
avoid creating a huge bottleneck on log retrieval. Remains to be seen
whether that's enough, my hunch is that it won't be enough for the
largest of instances (or tasks that have massive log streams, which
we'll probably have to process in a stream instead of concat-ing to a
big string)

## How I Tested These Changes
New assertions to existing tests.