Move pipes external metadata processing to its own layer. (#29311)
## Summary & Motivation
Pipes and airlift have kinda the same mandate here to use a
well-structured format for defining metadata in an external process,
that may not have access to the dagster package. So we can just move the
processing code to a shared place accessible to both, and make the
language more generic to instead be "external" metadata.

## How I Tested These Changes
Existing pipes metadata tests