[NO_SKIP=true] Run kitchen-sink on Docker queue
This started failing recently in ways that look like resource contention
on the instance. The agent itself is getting killed about 2/3 of the way
through.

This runs it with more resources. Something we previously did for a
similar variant of kitchen-sink tests:

https://github.com/dagster-io/dagster/pull/25487

If this build doesn't pass, I'll instead skip the test suite entirely.
Either way, we should carve out time to improve this suite's performance
and reliability.