RFC: Enable pytest-xdist in Buildkite
Most of our Buildkite runners have 2 vCPU. Some have 4. So we might be
able to get some pretty cheap improvements just by running more in
parallel. And we might find that there's a sufficient enough tradeoff
here that we can begin running larger Builkite runners.

I think we've made several half-hearted attempts at this in the past. My
guess is that we'll:

- unearth certain kinds of fixtures that cannot safely use this strategy
  (for example, ones that rely heavily on docker compose or already do
  their own crazy multi-process stuff). But I suspect once identified,
  we can split these into their own variant that don't parallelize.
- unearth non-determinism in test collection that prevents the initial
  distribution from happening. But I suspect once identified, we can
  pretty quickly burn this list down.
- make our logs kind of impossible to read. But they already are. We
  might be able to do something fun with Buildkite artifacts to log per
  test or something.

Anyway, putting this up to see what happens and for comment.

Some initial highlights on Buildkite with baseline vs. 2 vCPU:

Baseline: https://buildkite.com/dagster/dagster-dagster/builds/121620
2 vCPU: https://buildkite.com/dagster/dagster-dagster/builds/121622

- dagster cli tests: 17:08 -> 10:15
- dagster core tests: 15:23 -> 10:27
- dagster general tests: 8:13 -> 6:19
- dagster storage tests: 11:57 -> 6:45

With presumably more benefit to be had if we move these onto runners
with move vCPUs.

But of course, not everything is sunshine and roses yet.

If we generally like this direction, I'll spend some time generalizing
this across the test suite, playing whack-a-mole, and pricing out our
upsizing our Buildkite runners.