RFC: Fetch all code server data in DA before writing the initial tick cursor
Summary:
The way that DA attempts to ensure that each run is launched exactly once is as follows:

1). DA calculations figure out what runs to be requested, reserves a bunch of run IDs
2). A cursor is written with those run requests and run IDs
3) DA attempts to launch each run in succession, hitting the code server to generate job and execution plan snapshots for each one. It checks that a run does not already exist with the reserved run ID before launching.

If there is a failure during step 1, the tick can just retry from scrach - nothing has been written, so there's no risk of duplicated work.
If there is a failure during step 3, things are trickier. We don't want DA to get stuck in a loop where it can't proceed if the failure is permanent, but if it proceeds without launching everything it was supposed to, since the cursor is already written, work might get skipped.

The most common issue that causes a problem during step 3 is a change in the underlying workspace or code servers. For example, if a code location moves into an error state, the ideal behavior would be to not launch anything for that code location and not update the cursor and just wait, so that when the code location is back up and running again it can pick up where it left off.

But if an asset was renamed, for example, it's find to just ignore the cursor for the old asset, since it no longer exists at all.

In both of these cases, the best thing would be for the cursor to just not be written and for the tick to start over from scratch.

This PR attempts to bring us closer to that reality by doing all the reads that require code server access before writing the DA cursor, ensuring that as many failures as possible happen in step 1 where it is safe to retry and guarantee exactly-once semantics, vs. retrying in 3) where it is more dicery.

The main downside here is likely increased latency before any runs actually get submitted - since now all the snapshots need to be generated before a single run is actually created. The total amount of work that needs to happen is unchanged though, it's just the ordering of the steps that gets moved around.

This is an attempt to minimize the impact of race conditions that happen

> Insert changelog entry or delete this section.